+++ pg_lakehouse/src/fdw/cell.rs
