+++ pg_lakehouse/src/fdw/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod cell;
+mod datetime;
+mod local;
+mod s3;
